Allows attr 'callback' for formset-add directive.
This commit introduces the use of an attribute named 'callback'
for the formset-add directive, so that external functions can
be called right after a new form has been added to the formset.
A potential use case might consist of calling a function to
initialize an input field as a typeahead field. Very often such
a utility comes wrapped as a Jquery plugin, reason why the
external function is called with apply and receives the
ngDjangoFormset object.